+++ lib/src/models/view_configurations/view_configuration.dart
@@ -41,4 +41,4 @@ const defaultSnapIntervalMinutes = 10;
 const defaultShowEventTiles = true;
 const defaultInitialTimeOfDay = TimeOfDay(hour: 0, minute: 0);
 const defaultHeightPerMinute = 0.7;
-const defaultHorizontalPadding = 8.0;
+const defaultHorizontalPadding = EdgeInsets.only(left: 0, right: 4);